UInteger
The LSP specification defines the upper bound of the `uinteger`
type to 2^31 - 1 (Int.MAX_VALUE) instead of 2^32 - 1 (UInt.MAX_VALUE).
This is unlikely to be fixed because:
1. The LSP spec is following the DAP (Debug Adapter Protocol) spec.
2. The DAP spec won't fix the range to avoid making too many changes
   to the spec.

Note that the DAP spec does not have a `uinteger` type but instead
specifies the range of number values in the various interfaces.